+++ CHANGELOG.md
@@ -6,6 +6,9 @@ branch is the latest unreleased version.
 **Incompatibilities**
 - The RootDesc type is renamed to [Desc](https://github.com/Anaminus/rbxmk/blob/imperative/doc/types.md#user-content-desc).
 	- RootDesc.new is now [Desc.new](https://github.com/Anaminus/rbxmk/blob/imperative/doc/types.md#user-content-descnew)
+- The Optional functions under the [type library](https://github.com/Anaminus/rbxmk/blob/imperative/doc/libraries.md#user-content-types) are moved to constructors of the [Optional](https://github.com/Anaminus/rbxmk/blob/imperative/doc/types.md#user-content-optional) type.
+	- types.none is now [Optional.none](https://github.com/Anaminus/rbxmk/blob/imperative/doc/types.md#user-content-optionalnone).
+	- types.some is now [Optional.some](https://github.com/Anaminus/rbxmk/blob/imperative/doc/types.md#user-content-optionalsome).
 
 **Highlights:**
 - Add `plugin` dump format, which enables custom dump formats.